💡: 사소하게 계속 바꿔야할 게 있네요:(. 자료형 수정 및 디버깅용 모든 데이터 삭제 코드 추가

+++ SSENG/CoreData/data/HistoryCoreData/History+CoreDataProperties.swift
@@ -15,9 +15,9 @@ public extension History {
   }
 
   @NSManaged var userId: String? // 탑승했던 탑승자 아이디
-  @NSManaged var duration: Int // 총 이용 시간
+  @NSManaged var duration: Int16 // 총 이용 시간
   @NSManaged var startTime: String? // 탑승 시간
-  @NSManaged var type: Int // 킥보드 타입(1, 2)
+  @NSManaged var type: Int16 // 킥보드 타입(1, 2)
 }
 
 extension History: Identifiable {}
